requirements.txt
openai-agents>=0.0.8
python-dotenv>=1.0.0
rich>=13.5.0
pydantic>=2.0.0
